Extract message handling from EntryPreviewView
Make testing message handling easier without always having to set up
the whole view including seed HTML.
