--- entry_types/scrolled/package/spec/support/fakeParentWindow.js
@@ -1,5 +1,0 @@
-import { JSDOM } from 'jsdom';
-
-export function fakeParentWindow() {
-  Object.defineProperty(window, 'parent', {value: new JSDOM('').window });
-};
