+++ entry_types/scrolled/package/src/editor/views/EntryPreviewView.js
@@ -1,8 +1,7 @@
 import $ from 'jquery';
 import Marionette from 'backbone.marionette';
 import {cssModulesUtils} from 'pageflow/ui';
-import {watchCollections} from '../../entryState';
-import {InsertContentElementDialogView} from './InsertContentElementDialogView'
+import {PreviewMessageController} from '../controllers/PreviewMessageController'
 import styles from './EntryPreviewView.module.css'
 
 export const EntryPreviewView = Marionette.ItemView.extend({
@@ -14,89 +13,19 @@ export const EntryPreviewView = Marionette.ItemView.extend({
 
   ui: cssModulesUtils.ui(styles, 'iframe'),
 
-  initialize() {
-    this.messageListener = this.onMessage.bind(this);
-  },
-
   onShow() {
-    window.addEventListener('message', this.messageListener);
+    this.messageController = new PreviewMessageController({
+      entry: this.model,
+      editor: this.options.editor,
+      iframeWindow: this.ui.iframe[0].contentWindow
+    });
 
     inject(this.ui.iframe[0],
            unescape($('[data-template="iframe_seed"]').html()));
   },
 
   onClose() {
-    window.removeEventListener('message', this.messageListener);
-  },
-
-  onMessage(message) {
-    if (window.location.href.indexOf(message.origin) === 0) {
-      if (message.data.type === 'READY') {
-        const postMessage = message => {
-          this.ui.iframe[0].contentWindow.postMessage(message, window.location.origin);
-        };
-
-        watchCollections(this.model, {
-          dispatch: action => {
-            postMessage({type: 'ACTION', payload: action})
-          }
-        });
-
-        this.listenTo(this.model, 'scrollToSection', section =>
-          postMessage({
-            type: 'SCROLL_TO_SECTION',
-            payload: {
-              index: this.model.sections.indexOf(section)
-            }
-          })
-        );
-
-        this.listenTo(this.model, 'selectContentElement', contentElement =>
-          postMessage({
-            type: 'SELECT',
-            payload: {
-              id: contentElement.id,
-              type: 'contentElement'
-            }
-          })
-        );
-
-        this.listenTo(this.model, 'resetSelection', contentElement =>
-          postMessage({
-            type: 'SELECT',
-            payload: null
-          })
-        );
-      }
-      else if (message.data.type === 'CHANGE_SECTION') {
-        this.model.set('currentSectionIndex', message.data.payload.index);
-      }
-      else if (message.data.type === 'SELECTED') {
-        const {id} = message.data.payload;
-        const editor = this.options.editor;
-
-        if (id) {
-          editor.navigate(`/scrolled/content_elements/${id}`, {trigger: true})
-        }
-        else {
-          editor.navigate('/', {trigger: true})
-        }
-      }
-      else if (message.data.type === 'INSERT_CONTENT_ELEMENT') {
-        const {id, position, at} = message.data.payload;
-        const editor = this.options.editor;
-
-        InsertContentElementDialogView.show({
-          entry: this.model,
-          insertOptions: {position, id, at},
-          editor
-        });
-      }
-      else if (message.data.type === 'UPDATE_CONTENT_ELEMENT') {
-        const {id, configuration} = message.data.payload;
-        this.model.contentElements.get(id).configuration.set(configuration);
-      }
-    }
+    this.messageController.dispose();
   }
 });
